Add mantine ui design & routing

+++ nest/dark-meal-ai/src/recipes/recipes.controller.ts
@@ -1,15 +1,44 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+// MAKE RELATIONSHIP LOGIC
+// TO CREATE MEALS WITH ONE-TO-MANY RE
+
+
+import { 
+  Controller, 
+  Get, 
+  Post, 
+  Body, 
+  Patch, 
+  Param, 
+  Delete, 
+  UseGuards, 
+  Req, 
+  UnauthorizedException,
+  ForbiddenException
+} from '@nestjs/common';
 import { RecipesService } from './recipes.service';
 import { CreateRecipeDto } from './dto/create-recipe.dto';
 import { UpdateRecipeDto } from './dto/update-recipe.dto';
+import { OptionalJwtAuthGuard } from 'src/auth/optional-jwt.guard';
 
 @Controller('recipes')
 export class RecipesController {
   constructor(private readonly recipesService: RecipesService) {}
 
+  @UseGuards(OptionalJwtAuthGuard)
   @Post()
-  create(@Body() createRecipeDto: CreateRecipeDto) {
-    return this.recipesService.create(createRecipeDto);
+  create(
+    @Req() req,
+    @Body() createRecipeDto: CreateRecipeDto
+  ) {
+
+    if (!req.user) {
+      throw new UnauthorizedException("You need to authorize first!");
+    }
+
+    return this.recipesService.create({
+      ...createRecipeDto,
+      authorId: req.user.id,
+    });
   }
 
   @Get()
@@ -22,13 +51,32 @@ export class RecipesController {
     return this.recipesService.findOne(+id);
   }
 
+  @UseGuards(OptionalJwtAuthGuard)
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateRecipeDto: UpdateRecipeDto) {
+  update(
+    @Param('id') id: string, 
+    @Req() req,
+    @Body() updateRecipeDto: UpdateRecipeDto
+  ) {
+
+    if (!req.user) {
+      throw new UnauthorizedException("You need to authorize first!");
+    }
+
     return this.recipesService.update(+id, updateRecipeDto);
   }
 
+  @UseGuards(OptionalJwtAuthGuard)
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  remove(
+    @Param('id') id: string, 
+    @Req() req
+  ) {
+
+    if (!req.user) {
+      throw new UnauthorizedException("You need to authorize first!");
+    }
+
     return this.recipesService.remove(+id);
   }
 }
